+++ internal/provider/provider.go
@@ -16,7 +16,9 @@ import (
 
 	"github.com/streamkap-com/terraform-provider-streamkap/internal/api"
 	ds "github.com/streamkap-com/terraform-provider-streamkap/internal/datasource"
+	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/client_credential"
 	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/destination"
+	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/kafka_user"
 	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/pipeline"
 	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/source"
 	"github.com/streamkap-com/terraform-provider-streamkap/internal/resource/tag"
@@ -237,6 +239,7 @@ func (p *streamkapProvider) DataSources(_ context.Context) []func() datasource.D
 		ds.NewTopicsDataSource,
 		ds.NewTopicDataSource,
 		ds.NewTopicMetricsDataSource,
+		ds.NewRolesDataSource,
 	}
 }
 
@@ -302,5 +305,7 @@ func (p *streamkapProvider) Resources(_ context.Context) []func() resource.Resou
 		transform.NewRollupResource,
 		transform.NewFanOutResource,
 		transform.NewTopicRouterResource,
+		kafka_user.NewKafkaUserResource,
+		client_credential.NewClientCredentialResource,
 	}
 }